6-size.c

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-	printf("Size of char: %i bytes(s)\n", sizeof(char));
-	printf("Size of an int: %i byte(s)\n", sizeof(int));
-	printf("Size of long int: %i byte(s)\n", sizeof(long int));
-	printf("Size of long long int: %i byte(s)\n", sizeof(long long));
-	printf("Size of float is %i byte(s)\n", sizeof(float));
+	printf("Size of char: %d bytes(s)\n", (int) sizeof(char));
+	printf("Size of an int: %d byte(s)\n", (int) sizeof(int));
+	printf("Size of long int: %d byte(s)\n", (int) sizeof(long int));
+	printf("Size of long long int: %d byte(s)\n", (int) sizeof(long long));
+	printf("Size of float is %d byte(s)\n", (int) sizeof(float));
 	return (0);
 }
